character: belief/faction/group/world are now the string/names
Internally we still have the object $id as the relation, they're just
not visible any more.  $field now has the name of the related object.
$field can be modified by assigning to it (PUT/POST data).
The $field value is used to lookup the object which has that name and
we set the relation to it (by modifying the PUT/POST data so that it's
id is under $field_id).

If the $field is null or "" we lookup the object with name "-".

If no object is found in the case of belief/group or world, we create it.