bypass CakePHP on index() pages for performace
Outputting some index() pages took quite a long time due to the number of
items retrieved.  The characters listings as 3000+ entries took over 6
seconds to output on my RPi B v1.1.  Of which < 0.01 sec was spend on the
query.  The problem is in the php code, not the sql.

Digging showed that the ResultSet alone took 3 sec on it's ->toArray() call.
CakePHP is using alot of processing power to cast and convert the data to all
the nice data model objects.  But since they are not strictly needed for
rendering the index() as json, this is now replace by direct SQL queries which
gets directly transformed to the outputted dataformat.
Performance of the 6+ sec page is now < .5sec, win of a factor 10.

All the other pages are unchanged as there we want/need the strength of the
data model.  Performance is also less an issue there as the amount of data
being modified is small.