auditing: log previous entity-state to db
audits table contains everything needed to recreate an earlier
state of the entire database, or just specific entities.
(restoring, or even viewing is not yet implemented)

Saving an (existing) entity, results into writing the
entire previous/original state to the audits table.
When an entity is delete, also save the old state,
and write a second entry about when/who deleted it.

Lammies/print queue changes aren't logged to the audits table.