migrations: remove spells tables
Remove both 'spells' and 'characters_spells' database tables.
Removal of the code is in a seperate commit, because if we ever
wish to revert we will need to keep this migration file.

On migrate ("up()") the content of the tables is also written to
the 'history' table.  Each characters_spells row also get a
'removal' (data == null) entry.

The rollback ("down()") recreates the tables and tries to restore
their contents from what can be found in the 'history' table.