/auth/social/{,provider} for logging in via social
/auth/social
It returns a list of all the supported social login providers.
For each provider there's a 'url' and 'authUri' link, both need to be
customized by the front-end before they can be used.

The 'authUri' is used first.
Replace the STATE and CALLBACK strings in 'authUri':
- STATE is a random string used to prefent cross-site request forgery
- CALLBACK is where the user gets redirect to after login
Redirect the user to 'authUri' to start the login proces.

If login succesful, the user gets redirected to our CALLBACK location.
First we check that the 'state' matches with our expected value.
Next replace CODE and CALLBACK in the 'url' of the social provider:
- CODE is the string we got in the query parameter after the login
- CALLBACK must be the same as used in the 'authUri'
Perform a GET on the 'url' should yield a JWT that can be used for all
following interactions with the void api.
I.e. the user is logged in now.

The url of the front-end that is used in CALLBACK needs to be configured
as a know 'redirect_uri' at the third party / social login provider.

TODO: generate propper error responses instead of echo+die