remove unused item.attributes
The attributes/codes are no longer on the item card and the associated
skill has been deprecated.  So removing this unused functionality.

These item attributes were once used to hint players to the special
properties of the item.  Each attribute had a specific two letter code
that was put on the lammy card.  These that could then be decyphered
using the `recognize special object` skill.